Move filter classes to their own files
